requirements.txt
chainlit
langchain
openai
jira
boto3
atlassian-python-api
faiss-cpu
tiktoken
